去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@23710 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsVSAnnounce.cpp
@@ -64,7 +64,7 @@ namespace engine
       {
          next = id() ;
       }
-      else if ( MSG_CLS_BALLOT_RES == (UINT32)header->opCode )
+      else if ( MSG_CLS_BALLOT_RES == header->opCode )
       {
          const _MsgClsElectionRes *msg = ( const _MsgClsElectionRes * )
                                           header ;
