+++ SequoiaDB/engine/cls/clsVSVote.cpp
@@ -87,7 +87,7 @@ namespace engine
             }
          }
       }
-      else if ( MSG_CLS_BALLOT_RES == (UINT32)header->opCode )
+      else if ( MSG_CLS_BALLOT_RES == header->opCode )
       {
          const _MsgClsElectionRes *msg = ( const _MsgClsElectionRes * )
                                          header ;
